fix(popover): implement popover component, add ct tests IX-3942
